fix(vault): Make NegativeProfit error display balance
Currently the error message is unhelpful as it displays the old balance
before the loan occurred, and the required amount. It would be helpful
for errors if they displayed the balance the contract got to, so we can
see the difference between what was received as output, and what was
required.